Corrected Transition props enter & exit, fixed Slide props function

+++ src/main/scala/com/pangwarta/sjrmui/internal/Transition.scala
@@ -8,8 +8,8 @@ import scala.scalajs.js.|
 
 @js.native
 trait Transition extends js.Object {
-  var enter: Double = js.native
-  var exit: Double = js.native
+  var enter: Int = js.native
+  var exit: Int = js.native
 }
 
 object Transition {
